+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes are documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## 1.17.6 - Unreleased
+## 1.17.6 - 2019-10-18
 
 * Fixed [#636](https://github.com/NeoVintageous/NeoVintageous/issues/636): `cs"<div class="x">` should strip class from closing tag
 * Fixed [#636](https://github.com/NeoVintageous/NeoVintageous/issues/636): `ys{motion}<div class="x">` should strip class from closing tag